Corpus: Allow stopping after max_chars characters.
If the `max_chars` argument is supplied to Corpus(subdir=…), tokenizing
a file aborts after reading `max_chars` word characters.

This counts only characters that are part of a token, it adds 1 for
every token (representing the word separator), and it does not add the
word that does not fit into the character limit.

Rationale:

We want a character count instead of a word count to account for
corpora with different average word lengths. We don't just cut off after
n raw bytes to (a) ignore different formatting and (b) ignore encoding
issues.